# docs/pages/swaggers/swaggerMain.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MMGIS API",
    "description": "",
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "localhost:8888",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/login": {
      "post": {
        "description": "Log in to MMGIS",
        "parameters": [
          {
            "name": "master",
            "in": "body",
            "description": "Whether to use the master database instead of the mission relative one. (default: false)",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "mission",
            "in": "body",
            "description": "Mission name",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "body",
            "description": "Username",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User is logged in."
          }
        }
      }
    },
    "/signup": {
      "post": {
        "description": "Sign up for MMGIS",
        "parameters": [
          {
            "name": "master",
            "in": "body",
            "description": "Whether to use the master database instead of the mission relative one. (default: false)",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "mission",
            "in": "body",
            "description": "Mission name",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "body",
            "description": "Username",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "body",
            "description": "Email address",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User signed up and logged in."
          }
        }
      }
    },
...
  },
...
}